cria a configuracao do teste de integracao

+++ src/infra/db/typeorm/setup.js
@@ -1,15 +1,27 @@
+const { resolve } = require('path');
 const typeorm = require('typeorm');
 
-const typeormServer = new typeorm.DataSource({
-  type: 'sqlite',
-  database: 'db.sqlite',
-  synchronize: true,
-  dropSchema: true,
-  entities: [
-    require('./entities/Usuario.entity-typeorm'),
-    require('./entities/Livro.entity-typeorm'),
-    require('./entities/Emprestimo.entity-typeorm')
-  ]
-});
+let typeormServer;
+
+if (process.env.NODE_ENV === 'test') {
+  typeormServer = new typeorm.DataSource({
+    type: 'sqlite',
+    database: 'db.sqlite',
+    synchronize: true,
+    dropSchema: true,
+    entities: [resolve(__dirname, 'entities/*.entity-typeorm.js')]
+  });
+} else if (process.env.NODE_ENV === 'integration') {
+  typeormServer = new typeorm.DataSource({
+    type: 'postgres',
+    host: 'localhost',
+    database: 'biblioteca_test',
+    synchronize: true,
+    port: 5432,
+    username: 'postgres',
+    password: 'postgres',
+    entities: [resolve(__dirname, 'entities/*.entity-typeorm.js')]
+  });
+}
 
 module.exports = { typeormServer };
